livekit-agent/pyproject.toml
[project]
name = "bey-livekit-agent"
version = "0.1.0"
description = "Minimal avatar agent using the Beyond Presence API"
authors = [{ name = "Beyond Presence GmbH", email = "info@beyondpresence.ai" }]
requires-python = "~= 3.9"
dependencies = [
    "livekit-agents == 1.0.11",
    "livekit-plugins-bey @ git+https://github.com/niqodea/livekit-agents.git@bey-plugin-0.0.3#subdirectory=livekit-plugins/livekit-plugins-bey",
    "livekit-plugins-openai == 1.0.11",
    "livekit-plugins-silero == 1.0.11",
    "python-dotenv ~= 1.0",
]
